reduce dependencies for coreutils-single
Don't depend on libcrypto from the multicall binary in
the coreutils-single package, because that trades space for speed.

Don't depend on libgmp from the multicall binary in
the coreutils-single package, because that only benefits
factor and expr.  The large size of gmp is not seen as
an appropriate tradeoff for this functionality.

Note also there is reduced startup overhead for all tools
with these removed due to not linking with the shared libs.